Converted importer into a worker
