+++ impexp/src/main/AndroidManifest.xml
@@ -2,11 +2,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <application>
-        <service
-            android:name="com.adsamcik.tracker.impexp.importer.service.ImportService"
-            android:exported="false"
-            android:foregroundServiceType="specialUse"
-            tools:targetApi="q" />
-    </application>
 </manifest>
