+++ impexp/src/main/java/com/adsamcik/tracker/impexp/importer/DataImporter.kt
@@ -2,16 +2,21 @@ package com.adsamcik.tracker.impexp.importer
 
 import android.content.Context
 import android.net.Uri
+import androidx.work.OneTimeWorkRequestBuilder
+import androidx.work.WorkManager
+import androidx.work.workDataOf
 import com.adsamcik.tracker.impexp.importer.service.ImportService
+import com.adsamcik.tracker.impexp.importer.worker.ImportWorker
 import com.adsamcik.tracker.shared.base.extension.startForegroundService
 
 /**
  * Exposes import start to other packages.
  */
 object DataImporter {
-	fun import(context: Context, uri: Uri) {
-		context.startForegroundService<ImportService> {
-			putExtra(ImportService.ARG_FILE_URI, uri)
-		}
+	fun import(context: Context, fileUri: Uri) {
+		val workRequest = OneTimeWorkRequestBuilder<ImportWorker>()
+			.setInputData(workDataOf(ImportWorker.ARG_FILE_URI to fileUri.toString()))
+			.build()
+		WorkManager.getInstance(context).enqueue(workRequest)
 	}
 }
